feat(dasclaw_mcp): migrate 4 concrete transports from ironclaw (F3.2 phase 3 PR 6-β)
将 ironclaw `tools/mcp` 下 4 个 transport 实现按 verbatim 方式平移到
`crates/dasclaw_mcp/`：

- stdio_transport.rs (228 LOC) → child process stdin/stdout JSON-RPC
- unix_transport.rs  (234 LOC, #[cfg(unix)]) → Unix 域套接字 JSON-RPC
- process.rs         (206 LOC) → McpProcessManager 子进程生命周期
- http_transport.rs  (598 LOC) → Streamable HTTP / SSE，含 sanitize_error_body

dasclaw_mcp 这一侧仅有的编辑：
- import 路径机械替换：crate::tools::mcp::{protocol,session,transport,stdio_transport}
  → crate::{...}；crate::tools::tool::ToolError → dasclaw_tool::ToolError
- Cargo.toml 增加 tokio features process+net，reqwest features stream+rustls-tls-native-roots，
  加 futures = 0.3（http_transport SSE 用到 StreamExt + bytes_stream）
- transport::stream_transport_send 可见性从 PR 6-α 的临时 pub 收回到 pub(crate)
  （现在两个消费者 stdio/unix 已经同 crate，不再需要跨 crate）
- lib.rs 增加 4 个 pub mod + 公共类型 re-export

ironclaw 这一侧 4 个文件改为 11–13 行 pub use shim，保持
crate::tools::mcp::{stdio_transport,unix_transport,process,http_transport}::*
路径不变，外部消费者（extensions::manager、channels::web::server、
tools::builtin::extension_tools、tools::mcp::client、tools::mcp::factory）零改动。

不包含：factory.rs（依赖 McpClient）、client.rs（1475 LOC）——
留给 PR 6-γ（issue #665，待开）。

验证：
- cargo check -p dasclaw_mcp --tests：0 错 0 警
- cargo nextest run -p dasclaw_mcp：165/165 通过（PR 6-α 后 139/139 → +26 transport 测试）
- cargo check -p dasclaw --tests：0 错 0 警（仅保留 pre-existing main.rs 多 bin target 提示）
- cargo clippy --no-deps -p dasclaw_mcp --all-targets -- -D warnings：0 警告
- cargo fmt --all：无 diff
- python3.12 scripts/check_no_panics.py --base origin/xClaw：OK

Closes #664
Refs #661
